Add DateOfBirth and NINO to v3 teacher response
