+++ QualifiedTeachersApi/src/QualifiedTeachersApi/V3/Controllers/TeacherController.cs
@@ -12,6 +12,7 @@ using Swashbuckle.AspNetCore.Annotations;
 namespace QualifiedTeachersApi.V3.Controllers;
 
 [ApiController]
+[Route("teacher")]
 public class TeacherController : Controller
 {
     private readonly IMediator _mediator;
@@ -23,7 +24,6 @@ public class TeacherController : Controller
 
     [Authorize(AuthorizationPolicies.IdentityUserWithTrn)]
     [HttpGet]
-    [Route("teacher")]
     [SwaggerOperation(
         summary: "Get teacher details",
         description: "Gets the details of the currently authenticated teacher")]
@@ -53,26 +53,5 @@ public class TeacherController : Controller
         return Ok(response);
 
         IActionResult MissingOrInvalidTrn() => BadRequest();
-    }
-
-    [Authorize(AuthorizationPolicies.ApiKey)]
-    [HttpGet("teachers/{Trn}")]
-    [SwaggerOperation(
-        summary: "Get teacher details by TRN",
-        description: "Gets the details of the teacher corresponding to the given TRN")]
-    [ProducesResponseType(typeof(GetTeacherResponse), StatusCodes.Status200OK)]
-    [ProducesResponseType(StatusCodes.Status400BadRequest)]
-    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
-    [ProducesResponseType(StatusCodes.Status404NotFound)]
-    public async Task<IActionResult> Get([FromRoute] GetTeacherRequest request)
-    {
-        var response = await _mediator.Send(request);
-
-        if (response is null)
-        {
-            return NotFound();
-        }
-
-        return Ok(response);
     }
 }
