+++ QualifiedTeachersApi/src/QualifiedTeachersApi/V3/Controllers/TeachersController.cs
@@ -6,6 +6,7 @@ using Microsoft.AspNetCore.Mvc;
 using QualifiedTeachersApi.Filters;
 using QualifiedTeachersApi.Security;
 using QualifiedTeachersApi.V3.Requests;
+using QualifiedTeachersApi.V3.Responses;
 using Swashbuckle.AspNetCore.Annotations;
 
 namespace QualifiedTeachersApi.V3.Controllers;
@@ -20,6 +21,27 @@ public class TeachersController : ControllerBase
     public TeachersController(IMediator mediator)
     {
         _mediator = mediator;
+    }
+
+    [Authorize(AuthorizationPolicies.ApiKey)]
+    [HttpGet("{Trn}")]
+    [SwaggerOperation(
+        summary: "Get teacher details by TRN",
+        description: "Gets the details of the teacher corresponding to the given TRN")]
+    [ProducesResponseType(typeof(GetTeacherResponse), StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status400BadRequest)]
+    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    public async Task<IActionResult> Get([FromRoute] GetTeacherRequest request)
+    {
+        var response = await _mediator.Send(request);
+
+        if (response is null)
+        {
+            return NotFound();
+        }
+
+        return Ok(response);
     }
 
     [HttpPost("name-changes")]
