+++ QualifiedTeachersApi/src/QualifiedTeachersApi/V3/Handlers/GetTeacherHandler.cs
@@ -31,6 +31,8 @@ public class GetTeacherHandler : IRequestHandler<GetTeacherRequest, GetTeacherRe
                 Contact.Fields.FirstName,
                 Contact.Fields.MiddleName,
                 Contact.Fields.LastName,
+                Contact.Fields.BirthDate,
+                Contact.Fields.dfeta_NINumber,
                 Contact.Fields.dfeta_QTSDate,
                 Contact.Fields.dfeta_EYTSDate
             });
@@ -116,8 +118,10 @@ public class GetTeacherHandler : IRequestHandler<GetTeacherRequest, GetTeacherRe
         {
             Trn = request.Trn,
             FirstName = teacher.FirstName,
-            LastName = teacher.LastName,
             MiddleName = teacher.MiddleName,
+            LastName = teacher.LastName,
+            DateOfBirth = teacher.BirthDate.Value.ToDateOnly(),
+            NationalInsuranceNumber = teacher.dfeta_NINumber,
             Qts = MapQts(teacher.dfeta_QTSDate?.ToDateOnly()),
             Eyts = MapEyts(teacher.dfeta_EYTSDate?.ToDateOnly()),
             Induction = MapInduction(induction, inductionPeriods),
