+++ QualifiedTeachersApi/src/QualifiedTeachersApi/V3/Responses/GetTeacherResponse.cs
@@ -14,6 +14,8 @@ public record GetTeacherResponse
     [SwaggerSchema(Nullable = false)]
     public required string LastName { get; init; }
     public required string MiddleName { get; init; }
+    public required DateOnly DateOfBirth { get; init; }
+    public required string NationalInsuranceNumber { get; init; }
     public required GetTeacherResponseQts Qts { get; init; }
     public required GetTeacherResponseEyts Eyts { get; init; }
     public required GetTeacherResponseInduction Induction { get; init; }
