+++ QualifiedTeachersApi/tests/QualifiedTeachersApi.Tests/V3/GetTeacherTests.cs
@@ -97,6 +97,8 @@ public class GetTeacherTests : ApiTestBase
                 lastName = teacher.LastName,
                 middleName = teacher.MiddleName,
                 trn = trn,
+                dateOfBirth = teacher.BirthDate?.ToString("yyyy-MM-dd"),
+                nationalInsuranceNumber = teacher.dfeta_NINumber,
                 qts = new
                 {
                     awarded = teacher.dfeta_QTSDate?.ToString("yyyy-MM-dd"),
@@ -199,6 +201,8 @@ public class GetTeacherTests : ApiTestBase
         var firstName = Faker.Name.First();
         var lastName = Faker.Name.Last();
         var middleName = Faker.Name.Middle();
+        var dateOfBirth = Faker.Identification.DateOfBirth().ToDateOnly();
+        var nino = Faker.Identification.UkNationalInsuranceNumber();
 
         var qtsDate = new DateOnly(1997, 4, 23);
         var eytsDate = new DateOnly(1995, 5, 14);
@@ -208,8 +212,10 @@ public class GetTeacherTests : ApiTestBase
             Id = teacherId,
             dfeta_TRN = trn,
             FirstName = firstName,
-            LastName = lastName,
             MiddleName = middleName,
+            LastName = lastName,
+            BirthDate = dateOfBirth.ToDateTime(),
+            dfeta_NINumber = nino,
             dfeta_QTSDate = qtsDate.ToDateTime(),
             dfeta_EYTSDate = eytsDate.ToDateTime(),
         };
@@ -311,7 +317,6 @@ public class GetTeacherTests : ApiTestBase
 
         inductionPeriod.Attributes.Add($"appropriatebody.{Account.PrimaryIdAttribute}", new AliasedValue(Account.EntityLogicalName, Account.PrimaryIdAttribute, Guid.NewGuid()));
         inductionPeriod.Attributes.Add($"appropriatebody.{Account.Fields.Name}", new AliasedValue(Account.EntityLogicalName, Account.Fields.Name, inductionPeriodAppropriateBodyName));
-
 
         return new[]
         {
